feat(regulation): helper function to resolve variables json from business
