+++ app/domain/regulations_helper.py
@@ -1,0 +1,17 @@
+def resolve_variables(dict_var, business):
+    res_dict = {}
+    for key, val in dict_var.items():
+        if isinstance(val, dict):
+            for transport_type, val2 in val.items():
+                if transport_type != business.transport_type.name:
+                    continue
+                if isinstance(val2, dict):
+                    for business_type, val3 in val2.items():
+                        if business_type != business.business_type.name:
+                            continue
+                        res_dict[key] = val3
+                else:
+                    res_dict[key] = val2
+        else:
+            res_dict[key] = val
+    return res_dict
